requirements.txt
jupyter
matplotlib
pandas
pandas-stubs
scikit-learn
seaborn
google-search-results
python-dotenv
requests
